:art: chore: 迁移 panel 相关配置到 global 中
